+++ scripts/pandapower_to_pyomo_inspection.py
@@ -1,5 +1,4 @@
 # SPDX-FileCopyrightText: 2023-2026 Institute for High Voltage Equipment and Grids, Digitalization and Energy Economics (IAEW), RWTH Aachen University
-# SPDX-FileCopyrightText: 2023 Steffen Kortmann
 #
 # SPDX-License-Identifier: MIT
 
@@ -23,7 +22,7 @@ All power quantities are stored in p.u. on net.sn_mva (baseMVA).
 
 Institut für Elektrische Anlagen und Netze, Digitalisierung und
 Energiewirtschaft (IAEW)
-(c) 2023, Steffen Kortmann
+Author: Steffen Kortmann (2023)
 """
 
 import warnings